requirements.txt
setuptools~=46.0.0
requests~=2.23.0
python-dotenv~=0.13.0
pandas~=1.0.3
tdameritrade~=0.1.0
pytest~=5.4.1
mock~=4.0.2
selenium
webdriver-manager